update realtime dan sistem login

+++ app/Livewire/TodoList.php
@@ -2,9 +2,11 @@
 
 namespace App\Livewire;
 
-use App\Models\Todo;
+use Illuminate\Support\Facades\Auth;
+use Livewire\Attributes\Layout;
 use Livewire\Component;
 
+#[Layout('layouts.app')]
 class TodoList extends Component
 {
     public string $title = '';
@@ -21,30 +23,43 @@ class TodoList extends Component
             'title' => 'required|min:1|max:255',
         ]);
 
-        Todo::create(['title' => $this->title]);
+        $todo = Auth::user()->todos()->create([
+            'title' => $this->title,
+            'done'  => false,
+        ]);
+
         $this->title = '';
+        $this->dispatch('todo-added', id: $todo->id);
     }
 
-    public function toggle(Todo $todo)
+    public function toggle(int $id)
     {
+        $todo = Auth::user()->todos()->findOrFail($id);
         $todo->update(['done' => !$todo->done]);
+        $this->dispatch('todo-toggled', done: $todo->done);
     }
 
-    public function delete(Todo $todo)
+    public function delete(int $id)
     {
+        $todo = Auth::user()->todos()->find($id);
+
+        if (!$todo) return;
+
         $todo->delete();
+        $this->dispatch('todo-deleted');
     }
 
     public function changeBg(string $bg)
     {
         $this->bg = $bg;
         session(['bg' => $bg]);
+        $this->dispatch('bg-changed');
     }
 
     public function render()
     {
         return view('livewire.todo-list', [
-            'todos' => Todo::latest()->get(),
+            'todos' => Auth::user()->todos()->latest()->get(),
         ]);
     }
 }
